python/requirements.txt
numpy
opencv-python
pyserial
tkinter
